Refactored the code and added delay
Added a small delay to guarantee that the writing in the SD Card

+++ main.ino
@@ -2,6 +2,8 @@
 #include "SD_ADAPTER.h"
 
 int count = 0;
+int MAX_RECORDS = 10000;
+int USE_SDCARD = 1;
 
 void setup() {
   Serial.begin(9600);
@@ -10,21 +12,23 @@ void setup() {
   Serial.println("Initializing system...");
 
   // Initialize SD card
-  if (!begin_sd()) {
-    Serial.println("SD card initialization failed. Check connections.");
-    while (1);
-  }
+  if(USE_SDCARD){
+    if (!begin_sd()) {
+      Serial.println("SD card initialization failed. Check connections.");
+      while (1);
+    }
 
-  // Remove the previous file if it exists and create a new one
-  removeFile();
-  File dataFile = SD.open(filename, FILE_WRITE);
-  if (dataFile) {
-    Serial.println("CSV file created successfully!");
-    dataFile.println("Time,TiltX,TiltY,Status");
-    dataFile.close();
-  } else {
-    Serial.println("Error creating CSV file!");
-    while (1);
+    // Remove the previous file if it exists and create a new one
+    removeFile();
+    File dataFile = SD.open(filename, FILE_WRITE);
+    if (dataFile) {
+      Serial.println("CSV file created successfully!");
+      dataFile.println("Time,TiltX,TiltY,Status");
+      dataFile.close();
+    } else {
+      Serial.println("Error creating CSV file!");
+      while (1);
+    }
   }
 
   // Initialize IMU sensor
@@ -43,16 +47,13 @@ void loop() {
   // Calculate tilt angles and posture status
   data tiltData = getTilt();
 
-  // Save the data to SD card
-  writeData(tiltData.tiltX, tiltData.tiltY, tiltData.status);
+  if (USE_SDCARD) {
+    // Save the data to SD card
+    writeData(tiltData.tiltX, tiltData.tiltY, tiltData.status);
+
+    delay(1000);
+  }
 
   printData();
 
-  delay(500);
-
-  count++;
-  if (count >= 10000) {
-    Serial.println("Finalizing the data storage process.");
-    while (1);
-  }
 }
